Add VGRC-VGOODRCF405DJI targets config
Hi Mikeller :
For the  safety about default with Dshot 600
We sell this controller with our latest BLHeli_32 ESC and our ESC default also with Dshot600, at the same time, we will write and enhance this safety instruction in our FC manual. We will take responsibility for any safety issue if there. tks.